Refactor #28: 게시글 작성 컴포넌트 폴더 이름 Discussion 기능에 맞게 변경
